library/requirements.txt
appnope==0.1.0
asgiref==3.3.0
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==3.6
decorator==4.4.2
dj-database-url==0.5.0
Django==3.1.3
django-crispy-forms==1.9.2
django-rest-swagger==2.2.0
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
ipdb==0.13.4
ipython==7.16.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.17.2
Jinja2==2.11.2
MarkupSafe==1.1.1
openapi-codec==1.3.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.8
psycopg2==2.8.6
psycopg2-binary==2.8.6
ptyprocess==0.6.0
Pygments==2.7.2
pytz==2020.4
PyYAML==5.3.1
requests==2.24.0
simplejson==3.17.2
sqlparse==0.4.1
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.11
wcwidth==0.2.5
whitenoise==5.2.0
